Subscription: add "list all" endpoint
For consistency, `Customer::Subscription` now inherits from `Subscription`,
similar to `Payment` and `Customer::Payment`.
