+++ lib/mollie.rb
@@ -16,8 +16,6 @@ require 'mollie/amount'
 require 'mollie/chargeback'
 require 'mollie/client'
 require 'mollie/customer'
-require 'mollie/customer/mandate'
-require 'mollie/customer/subscription'
 require 'mollie/invoice'
 require 'mollie/list'
 require 'mollie/method'
@@ -28,8 +26,11 @@ require 'mollie/permission'
 require 'mollie/profile'
 require 'mollie/refund'
 require 'mollie/settlement'
+require 'mollie/subscription'
 
+require 'mollie/customer/mandate'
 require 'mollie/customer/payment'
+require 'mollie/customer/subscription'
 require 'mollie/onboarding'
 require 'mollie/order/line'
 require 'mollie/order/refund'
